WIP - start extracting a common httpclient
