+++ lib/wallaby/experimental/selenium/webdriver_client.ex
@@ -314,104 +314,6 @@ defmodule Wallaby.Experimental.Selenium.WebdriverClient do
       do: value
   end
 
-  @type request_opts :: {:encode_json, boolean}
-
-  @doc """
-  Low-level function that sends a request to the webdriver API and parses the
-  response.
-  """
-  @spec request(http_method, url, map | String.t, [request_opts]) ::
-    {:ok, any} | {:error, :stale_reference | :invalid_selector}
-  def request(method, url, params \\ %{}, opts \\ [])
-  def request(method, url, params, _opts) when map_size(params) == 0 do
-    make_request(method, url, "")
-  end
-  def request(method, url, params, [{:encode_json, false} | _]) do
-    make_request(method, url, params)
-  end
-  def request(method, url, params, _opts) do
-    make_request(method, url, Poison.encode!(params))
-  end
-
-  @doc """
-  Low-level function that sends a request to the webdriver API and
-  raises an exception if an error occurs.
-  """
-  @spec request!(http_method, url) :: any
-  def request!(method, url) do
-    make_request!(method, url, "")
-  end
-
-  defp make_request(method, url, body) do
-    HTTPoison.request(method, url, body, headers(), request_opts())
-    |> handle_response
-    |> case do
-         {:error, :httpoison} ->
-           make_request(method, url, body)
-         result ->
-           result
-    end
-  end
-
-  defp make_request!(method, url, body) do
-    case make_request(method, url, body) do
-      {:ok, resp} ->
-        resp
-
-      {:error, :stale_reference} ->
-        raise Wallaby.StaleReferenceException
-
-      {:error, :invalid_selector} ->
-        raise Wallaby.InvalidSelector, Poison.decode!(body)
-
-      {:error, e} ->
-        raise "There was an error calling: #{url} -> #{e.reason}"
-    end
-  end
-
-  defp request_opts do
-    Application.get_env(:wallaby, :hackney_options, [])
-  end
-
-  defp headers do
-    [{"Accept", "application/json"},
-      {"Content-Type", "application/json"}]
-  end
-
-  defp handle_response({:error, %HTTPoison.Error{}}) do
-    {:error, :httpoison}
-  end
-  defp handle_response({:ok, %HTTPoison.Response{status_code: 204}}) do
-    {:ok, %{"value" => nil}}
-  end
-  defp handle_response({:ok, %HTTPoison.Response{body: body}}) do
-    with {:ok, decoded} <- Poison.decode(body),
-          {:ok, validated} <- check_for_response_errors(decoded),
-          do: {:ok, validated}
-  end
-  defp handle_response({:error, reason}), do: {:error, reason}
-
-
-  defp check_for_response_errors(response) do
-    case Map.get(response, "value") do
-      %{"class" => "org.openqa.selenium.StaleElementReferenceException"} ->
-        {:error, :stale_reference}
-      %{"message" => "stale element reference" <> _} ->
-        {:error, :stale_reference}
-      %{"class" => "org.openqa.selenium.InvalidSelectorException"} ->
-        {:error, :invalid_selector}
-      _ ->
-        {:ok, response}
-    end
-  end
-
-  defp to_params({:xpath, xpath}) do
-    %{using: "xpath", value: xpath}
-  end
-  defp to_params({:css, css}) do
-    %{using: "css selector", value: css}
-  end
-
   @spec cast_as_element(Session.t | Element.t, map) :: Element.t
   defp cast_as_element(parent, %{"ELEMENT" => id}) do
     %Wallaby.Element{
